- stage 설정 추가

+++ daema-api/src/main/java/com/daema/rest/base/web/MemberController.java
@@ -147,10 +147,8 @@ public class MemberController {
             String profile = PropertiesValue.profilesActive;
 
             if(profile != null &&
-                    (!"prod".equals(profile) ||
-                            !"stag".equals(profile)
-                    )
-            ) {
+                    (!"prod".equals(profile) &&
+                            !"stag".equals(profile))) {
                 redisUtil.deleteData("localUserid");
             }
